refactor: Renamed PostersTitle to OverviewTitle 🚚
